Move ProcessingContext#report to RetryWithToleranceOperator to avoid keeping state in ProcessingContext
